Fix sync offset: it shall works on the delta, not the absolute drift that will always diverge. Refactor timer class.
