+++ lib/CMakeLists.txt
@@ -8,6 +8,7 @@ set(KICKCAT_SOURCES
   ${CMAKE_CURRENT_SOURCE_DIR}/src/SIIParser.cc
 
   ${CMAKE_CURRENT_SOURCE_DIR}/src/OS/Time.cc
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/OS/Timer.cc
 
   ${CMAKE_CURRENT_SOURCE_DIR}/src/CoE/OD.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/src/CoE/protocol.cc
