+++ lib/include/kickcat/OS/Timer.h
@@ -1,12 +1,9 @@
 #ifndef KICKCAT_OS_TIMER_H
 #define KICKCAT_OS_TIMER_H
 
-#include <string_view>
 #include <system_error>
 
 #include "kickcat/OS/Time.h"
-#include "kickcat/OS/ConditionVariable.h"
-#include "kickcat/OS/Mutex.h"
 
 namespace kickcat
 {
@@ -17,19 +14,17 @@ namespace kickcat
         Timer(Timer const &) = delete;
         void operator=(Timer const &) = delete;
 
-        /// \param  name     Name of the timer (logging)
         /// \param  period   Interval between two timer firing
         Timer(nanoseconds period);
         ~Timer() = default;
 
-        /// Start the timer.
-        void start(nanoseconds sync_point = since_epoch());
+        /// \brief  Initialize (or restart) the timer.
+        /// \details Aligns the next deadline to sync_point + N*period so that it lies in the future.
+        ///          Resets the drift compensation state. Safe to call multiple times.
+        void init(nanoseconds sync_point = since_epoch());
 
-        ///  Stop the timer.
-        void stop();
-
-        ///  Get timer status
-        bool is_stopped() const;
+        /// \deprecated Use init() instead.
+        void start(nanoseconds sync_point = since_epoch()) { init(sync_point); }
 
         ///  Change the timer values to the new provided values and reset it.
         ///  This does not take effect before the timer is restarted.
@@ -42,27 +37,22 @@ namespace kickcat
         /// Wait until the next timer tick (blocking call)
         std::error_code wait_next_tick();
 
-        /// \brief  Adjust the timer phase to track an external time reference (e.g. EtherCAT DC).
+        /// \brief  Match the timer's firing rate to an external time reference (e.g. EtherCAT DC).
         /// \details Call this every cycle before wait_next_tick() with bus.dcMasterOffset().
-        ///          A positive offset means the master is ahead of the reference: the timer slows down.
-        ///          Uses a first-order IIR filter to smooth the correction.
-        ///          Pass 0ns when no reference is available (no-op when filtered offset is zero).
+        ///          The measured offset grows linearly at the clock drift rate: the timer
+        ///          tracks the per-cycle delta and uses it as a period correction, so the
+        ///          timer fires at the same real-time rate as the reference clock.
+        ///          Do not call this when no drift compensation is desired.
         void apply_offset(nanoseconds raw_offset);
 
-    protected:
     private:
-        static constexpr int64_t OFFSET_FILTER_DEPTH      = 256;
-        static constexpr int64_t OFFSET_CORRECTION_DIVISOR = 16;
+        static constexpr int64_t DRIFT_FILTER_DEPTH = 256;
 
-        std::string name_;
         nanoseconds period_;
         nanoseconds next_deadline_{};
         nanoseconds last_wakeup_{};
-        nanoseconds filtered_offset_{0ns};
-
-        Mutex mutex_{};
-        ConditionVariable stop_{};
-        bool is_stopped_{true};
+        nanoseconds last_raw_offset_{0ns};
+        nanoseconds filtered_drift_{0ns};
     };
 }
